Cleaned up functional version of data_processing_v2, labelling it as beta, and removed explicit Any
