+++ tools/data_processing_v2_beta/main.py
@@ -1,7 +1,10 @@
+# Data Processing v2 (beta) - Waterloo Rocketry
 
-from daq_processing.daq_processing import DAQDataProcessor
+from processors import DAQDataProcessor
 
-def main():
+
+# TODO: Make this an actual app and not a script (and maybe GUI?)
+def main() -> None:
     with open("2025_03_01-02_14_03_AM.log", "rb") as file:
         processor = DAQDataProcessor(file, "DAQ/Fake")
         processor.process()
