+++ tools/data_processing_v2_beta/processors/__init__.py
@@ -1,0 +1,4 @@
+# pyright: strict
+from .daq_processing import DAQDataProcessor, DAQDataStructure
+
+__all__ = ["DAQDataProcessor", "DAQDataStructure"]
